// app/(auth)/register/types.ts
export type ClientForm = {
  name: string;
  email: string;
  password: string;
  city: string;
  country: string;
};

export type GPForm = {
  name: string;
  email: string;
  password: string;
  city: string;
  destination: string;
  departureDate: string;
  availableKg: string;
  description: string;
};